Fix types and cast in some utils

+++ src/utils/hash/MD5.cpp
@@ -319,7 +319,7 @@ static void md5_init(md5_state_t *pms)
     pms->abcd[3] = 0x10325476;
 }
 
-static void md5_append(md5_state_t *pms, const md5_byte_t *data, int nbytes)
+static void md5_append(md5_state_t* pms, const md5_byte_t* data, size_t nbytes)
 {
 	if (!data || nbytes < 0)
 	{
@@ -332,7 +332,7 @@ static void md5_append(md5_state_t *pms, const md5_byte_t *data, int nbytes)
 	}
 
 	const md5_byte_t *p = data;
-    int left = nbytes;
+	size_t left = nbytes;
     int offset = (pms->count[0] >> 3) & 63;
     md5_word_t nbits = (md5_word_t)(nbytes << 3);
 
@@ -347,7 +347,7 @@ static void md5_append(md5_state_t *pms, const md5_byte_t *data, int nbytes)
     /* Process an initial partial block. */
     if (offset)
     {
-		int copy = (offset + nbytes > 64 ? 64 - offset : nbytes);
+		size_t copy = (offset + nbytes > 64 ? (64u - offset) : nbytes);
 
 		memcpy(pms->buf + offset, p, copy);
